# dev_core/api/openapi.yaml
openapi: 3.0.3
info:
  title: InduFrame Control Plane API
  version: 0.1.0
  description: dev_core 面向 dev_ide 的控制面契约。
servers:
  - url: /api/v1
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [project]
      summary: listProjects
      description: 稳定按updated_at DESC、id DESC分页。每个data.list项目附deploymentSummary，由当前页项目ID一次租户隔离批量查询；包含deploymentCount、environmentCount、operationInProgress、updatedAt、primarySelection及primaryDeployment。仅deploymentCount=1时primarySelection=unique并返回primaryDeployment；0为none，多部署为multiple且primaryDeployment=null，不以最近/默认环境部署代表全部。primary的mode为development|production，observedStatus沿正式部署状态，不新增deploying枚举。pending run、删除请求，或非失败/非异常部署中未失败运行引擎的desiredGeneration大于observedGeneration时，operationInProgress为true；updating保留代次细分。失败/异常部署、失败服务及反向代次不持续阻塞异常处理。placements/services是现有槽位预填，不替代当前制品的required引擎判断。摘要查询失败则整个列表失败，不返回伪造0。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      operationId: createProject
      tags: [project]
      summary: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/tags:
    get:
      operationId: listProjectTags
      tags: [project]
      summary: listProjectTags
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      operationId: createProjectTag
      tags: [project]
      summary: createProjectTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamedResourceRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/tags/{tagId}:
    put:
      operationId: updateProjectTag
      tags: [project]
      summary: updateProjectTag
      parameters:
        - name: tagId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamedResourceRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      operationId: deleteProjectTag
      tags: [project]
      summary: deleteProjectTag
      parameters:
        - name: tagId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/groups:
    get:
      operationId: listProjectGroups
      tags: [project]
      summary: listProjectGroups
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      operationId: createProjectGroup
      tags: [project]
      summary: createProjectGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamedResourceRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/groups/{groupId}:
    put:
      operationId: updateProjectGroup
      tags: [project]
      summary: updateProjectGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamedResourceRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      operationId: deleteProjectGroup
      tags: [project]
      summary: deleteProjectGroup
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}/tags:
    put:
      operationId: replaceProjectTags
      tags: [project]
      summary: replaceProjectTags
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceIDsRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}/group:
    put:
      operationId: setProjectGroup
      tags: [project]
      summary: setProjectGroup
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetProjectGroupRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}:
    put:
      operationId: updateProject
      tags: [project]
      summary: updateProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      operationId: deleteProject
      tags: [project]
      summary: deleteProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectDeleteRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}/authoring-context:
    get:
      operationId: getProjectAuthoringContext
      tags: [project]
      summary: getProjectAuthoringContext
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 成功, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
        default: { description: 失败, content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } } }
  /projects/{projectId}/delete-impact:
    get:
      operationId: getProjectDeleteImpact
      tags: [project]
      summary: getProjectDeleteImpact
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}/operations/{operation}:
    post:
      operationId: operateProject
      tags: [project]
      summary: operateProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: operation
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectOperationRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/{projectId}/export:
    get:
      operationId: exportProject
      tags: [project]
      summary: exportProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /projects/import:
    post:
      operationId: importProject
      tags: [project]
      summary: importProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectImportRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: NodeAgentToken
  schemas:
...
    ProjectCreateRequest:
      type: object
      required: [name]
      properties:
        template: { type: string, enum: [demo-shell], description: 创建组织内独立的工程模板外壳，不初始化内部示例；每组织最多一份 }
        name: { type: string }
        description: { type: string }
    ProjectUpdateRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
    ProjectDeleteRequest:
      type: object
      properties:
        confirmation: { type: string }
    ProjectOperationRequest:
      type: object
      properties:
        operation: { type: string }
...
    ProjectImportRequest:
      type: object
      properties:
        files: { type: object, additionalProperties: { type: string } }
    NamedResourceRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
    ReplaceIDsRequest:
      type: object
      properties:
        ids: { type: array, items: { type: string, format: uuid } }
        roleIds: { type: array, items: { type: string, format: uuid } }
        tagIds: { type: array, items: { type: string, format: uuid } }
    SetProjectGroupRequest:
      type: object
      properties:
        groupId: { type: string, format: uuid, nullable: true }
...
    ApiResponse:
      type: object
      required: [code, msg, data, reqId]
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          nullable: true
        reqId:
          type: string
security:
  - bearerAuth: []